create platform mods for syscalls and stub for windows
